requirements.txt
Flask==2.0.2
imutils==0.5.4
numpy==1.19.5
opencv_contrib_python==4.5.3.56
Pillow==8.4.0
torch==1.9.1
torchvision==0.10.1
